add Command struct (refactoring)
